+++ R/vein_notes.R
@@ -1,5 +1,4 @@
-#' vein_notes for writting technical notes about the inventory
-#'
+#' @title  Notes with sysinfo
 #' @description \code{\link{vein_notes}} creates aa text file '.txt' for
 #' writting technical notes about this emissions inventory
 #'
@@ -148,7 +147,7 @@ vein_notes <- function (notes,
   cat("========================================\n") # 40
   cat("\n\n\nThanks for using VEIN\n")
   sink()
-  cat("File at:", file, "\n")
-  return(file)
+  message("File at:", file, "\n")
+  # return(file)
 }