+++ projects/brazil_bu/scripts/net.R
@@ -1,17 +1,26 @@
 library(sf)
-# Edita para tuas necessidades
-cat("\nNames: ", names(net), "\n")
+# Edit
+
+switch (language,
+        "portuguese" = cat("\nNomes: ", names(net), "\n"),
+        "english" = cat("\nNames: ", names(net), "\n"),
+        "chinese" = cat("\n名字: ", names(net), "\n"),
+        "spanish" = cat("\nNombres: ", names(net), "\n"))
+
 net <- st_transform(net, crs)
-# A composição veicular da CETESB tem 28 tipos de veiculos
+# A composição veicular da CETESB tem 28 tipos de categories
 # Em teoría, poderiamos ter 28 fluxos ou mais na rede net
-for(i in seq_along(veiculos)) {
-  png(filename =  paste0("images/NET_", veiculos[i],".png"),
+
+# The number of traffic flow vehicles can be the same from vehicular composition
+#
+for(i in seq_along(categories)) {
+  png(filename =  paste0("images/NET_", categories[i],".png"),
       width = 2300, height = 1500, units = "px", pointsize = 12,
       bg = "white",  res = 300)
-  plot(net[veiculos[i]], 
+  plot(net[categories[i]], 
        axes = T, 
        pal = cpt(colorRampPalette = T, rev = T), 
-       main = paste0(tit, ": ", veiculos[i]))
+       main = paste0(tit, ": ", categories[i]))
   dev.off()
 }
 
@@ -19,12 +28,26 @@ net$lkm <- net$lkm*units::as_units("km")
 
 saveRDS(net, 'network/net.rds')
 
-message(paste0("Arquivos em ", getwd(), "/net\n"))
-message(paste0("Figuras em ", getwd(), "/images\n"))
-cat("Limpando... \n")
+switch (language,
+        "portuguese" = message("\nArquivos em: /net:"),
+        "english" = message("\nFiles in: /net"),
+        "chinese" = message("\n文件位于: /net"),
+        "spanish" = message("\nArchivos en: /net"))
+
+switch (language,
+        "portuguese" = message("Figuras em: /images:"),
+        "english" = message("Figures in: /images"),
+        "chinese" = message("中的数字: /images"),
+        "spanish" = message("Figuras en: /images"))
+
+switch (language,
+        "portuguese" = message("Limpando..."),
+        "english" = message("Cleaning..."),
+        "chinese" = message("清洁用品..."),
+        "spanish" = message("Limpiando..."))
 
 suppressWarnings(
-  rm(i, tit, net, veiculos, crs)
+  rm(i, tit, net, categories, crs)
   
 )
 gc()