+++ NEWS.md
@@ -7,6 +7,7 @@ NEWS
 - Added projet "emislacovid" (2020-07-11)
 - Improved colplot (2020-07-17)
 - Added scaled brazilian emission factors with tunnel measurements (2020-07-21) and speciations.
+- Remove evaporative emission factors of flex motorcycles for year 2008.
 
 ## vein 0.8.9  (Release date: 2020-06-11)